requirements/local.txt
asgiref==3.8.1
Django==5.0.7
django-model-utils==4.5.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
typing_extensions==4.12.2
Unipath==1.1